// src/services/interfaces.ts
export interface IHist {
  id: string;
  tipo: string;
  veiculo: string;
  driver: {name: string};
  cliente: string;
  obra: {nome: string};
  closed_at: string;
  created_at: string;
  costs: [{value:string},{value:string},{value:string}]
  custo_cliente: string;
  custo_empresa: string;
  custo_motorista: string;
}